Moves getPatientAppointments to a new PatientAppointmentController
